Add Cygwin reference to Windows options
To run openssl on Windows (https://pantheon.io/docs/enable-https/#generate-rsa-key-and-csr), the Docs reference Cygwin with a link on how to install (https://pantheon.io/docs/cygwin-windows/). On that page, it directly mentions using Terminus as a benefit,  yet the Terminus installation page only mentions Git Bash and BoUoW (Bash on Ubuntu on Windows).